Add auth

+++ src/main/java/com/github/enteraname74/musik/controller/MusicInformationController.java
@@ -1,13 +1,14 @@
 package com.github.enteraname74.musik.controller;
 
-import com.github.enteraname74.musik.domain.model.Music;
+import com.github.enteraname74.musik.controller.utils.ControllerUtils;
+import com.github.enteraname74.musik.domain.service.AuthService;
 import com.github.enteraname74.musik.domain.service.MusicInformationService;
 import com.github.enteraname74.musik.domain.utils.ServiceResult;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 /**
  * Controller for managing API calls related to musics.
@@ -16,12 +17,14 @@ import java.util.List;
 @RequestMapping("/music/information")
 public class MusicInformationController {
     private final MusicInformationService musicInformationService;
+    private final AuthService authService;
 
     @Autowired
     public MusicInformationController(
-            MusicInformationService musicInformationService
+            MusicInformationService musicInformationService, AuthService authService
     ) {
         this.musicInformationService = musicInformationService;
+        this.authService = authService;
     }
 
     /**
@@ -30,8 +33,14 @@ public class MusicInformationController {
      * @return a list containing all Musics or an empty list if there is an error.
      */
     @GetMapping("/all")
-    List<Music> getAll() {
-        return musicInformationService.getAll();
+    ResponseEntity<?> getAll(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token
+    ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
+        return new ResponseEntity<>(
+                musicInformationService.getAll(),
+                HttpStatus.OK
+        );
     }
 
     /**
@@ -42,8 +51,10 @@ public class MusicInformationController {
      */
     @GetMapping("/{id}")
     ResponseEntity<?> get(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token,
             @PathVariable String id
     ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
         ServiceResult<?> result = musicInformationService.getById(id);
 
         return new ResponseEntity<>(result.getResult(), result.getHttpStatus());
@@ -58,8 +69,10 @@ public class MusicInformationController {
      */
     @DeleteMapping("/{id}")
     ResponseEntity<?> delete(
+            @RequestHeader(name = HttpHeaders.AUTHORIZATION) String token,
             @PathVariable String id
     ) {
+        if (!authService.isUserAuthenticated(token)) return ControllerUtils.UNAUTHORIZED_RESPONSE;
         ServiceResult<?> result = musicInformationService.deleteById(id);
 
         return new ResponseEntity<>(result.getResult(), result.getHttpStatus());
